improved

+++ scripts/create-booteable-netinstall.sh
@@ -4,6 +4,24 @@ ISO=$1
 DEV=$2
 DIST=trusty
 
+function usage {
+  echo "create-booteable-netinstall.sh [iso_image] [device]"
+}
+
+if [ $(test -f ${ISO}) == 0 ]
+then
+  echo "${ISO}: is not a regular file"
+  usage
+  exit 1
+fi
+
+if [ $(test -b ${DEV}) == 0 ]
+then
+  echo "${ISO}: is not a block device"
+  usage
+  exit 1
+fi
+
 if [[ $(/usr/bin/id -u) -ne 0 ]]; then
     echo "This script must be run with root privilege. Trying to sudo command."
     exit -1
@@ -54,6 +72,8 @@ append initrd=/install/initrd.gz vga=normal auto file=/cdrom/preseed/openstack-b
 
 EOF
 
+mkdir -p $BUILD/boot/grub
+
 cat > $BUILD/boot/grub/grub.cfg << EOF
 
 if loadfont /boot/grub/font.pf2 ; then
@@ -87,22 +107,31 @@ echo "ISO Done!"
 
 echo "command parameters: " $1
 if [ -n "$DEV" ]; then
+  rmdir $TMPMNT 2>/dev/null
 	echo "Creating USB booteable...."
 	MNTTMP=$(mktemp -d)
 	umount ${DEV}* > /dev/null 2>&1
-	echo "Erasing USB stick ..."
-	dd if=/dev/zero of=$DEV bs=512 count=1 conv=notrunc
-	echo "Creating partitioning and filesystem"
-	install-mbr --force $DEV
-	parted -s $DEV mkpart primary 1 3500
-	parted -s $DEV set 1 boot on
-	mkfs.vfat ${DEV}1
+
+  if [ 1 == 0 ]
+  then
+	  echo "Erasing USB stick ..."
+	  dd if=/dev/zero of=$DEV bs=512 count=1 conv=notrunc
+	  echo "Creating partitioning and filesystem"
+	  install-mbr --force $DEV
+	  parted -s $DEV mkpart primary 1 3500
+	  parted -s $DEV set 1 boot on
+    DEV=${DEV}1
+	  mkfs.vfat -I -n openstack-builder ${DEV}
+  else
+    DEV=${DEV}1
+  fi
+
 	echo "Copying files ..."
-	mount ${DEV}1 $MNTTMP
+	mount ${DEV} $MNTTMP
 	rsync -va ${BUILD}/ ${MNTTMP}/
 	cp $BUILD/isolinux/isolinux.cfg ${MNTTMP}/syslinux.cfg
-	syslinux -i ${DEV}1
-	umount ${DEV}1
+	syslinux -i ${DEV}
+	umount ${DEV}
 	echo "USB Done!"
 fi
